fix: 1.changed folder structure 2.hide form 3.like&cit action 4.action for send empty fields in login
