+++ src/services/render/_commentPageRenderer.js
@@ -1,12 +1,14 @@
-import { commentsListenersInit } from '../../../events/commentListeners.js';
+import { commentsListenersInit } from '../../events/commentListeners.js';
 
 export function commentsPageRenderer() {
   const container = document.querySelector('#app');
   const hasName = 'name' in localStorage;
+  const isLogin = 'token' in localStorage;
+
   container.innerHTML = `      
     <div class="container" id="container-comments">
         <ul class="comments"></ul>
-        <div class="add-form" id="add-comment-form">
+        <div class="add-form ${isLogin ? '' : '-add-form-hidden'}" id="add-comment-form">
           <input
             type="text"
             class="add-form-name"
