+++ src/utils/createLoaderText.js
