+++ src/utils/xssValidate.js
